Rename files or components under `js/src/components` for better consistency in naming conventions.
