+++ js/src/components/free-listings/configure-product-listings/shipping-time/shipping-time-setup/add-time-button/add-time-modal/index.js
@@ -14,7 +14,7 @@ import AppModal from '.~/components/app-modal';
 import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import SupportedCountrySelect from '.~/components/supported-country-select';
 import validateShippingTimeGroup from '.~/utils/validateShippingTimeGroup';
-import MinMaxShippingTimes from '../../min-max-inputs';
+import MinMaxShippingTimes from '../../min-max-shipping-times';
 
 /**
  * Form to add a new time for selected country(-ies).
