+++ js/src/components/free-listings/configure-product-listings/shipping-time/shipping-time-setup/countries-time-input/index.test.js
@@ -54,7 +54,8 @@ describe( 'CountriesTimeInput', () => {
 			expect( getByDisplayValue( '' ) ).toBeInTheDocument();
 		} );
 	} );
-	describe( 'Test Stepper', () => {
+
+	describe( 'Test TimeStepper', () => {
 		it( 'Should call onChange when increasing an decreasing the days', async () => {
 			const onChange = jest.fn();
 			const { queryAllByRole } = render(
