+++ js/src/components/free-listings/configure-product-listings/shipping-time/shipping-time-setup/index.js
@@ -9,7 +9,7 @@ import { __ } from '@wordpress/i18n';
 import { useAdaptiveFormContext } from '.~/components/adaptive-form';
 import Section from '.~/wcdl/section';
 import AppSpinner from '.~/components/app-spinner';
-import ShippingCountriesForm from './countries-form';
+import ShippingCountriesForm from './shipping-countries-form';
 
 /**
  * Form control to edit shipping rate settings.
