+++ js/src/components/free-listings/configure-product-listings/shipping-time/shipping-time-setup/time-stepper/index.js
@@ -12,7 +12,7 @@ import { plus, reset } from '@wordpress/icons';
 import AppInputNumberControl from '.~/components/app-input-number-control';
 import './index.scss';
 
-const Stepper = ( {
+const TimeStepper = ( {
 	step = 1,
 	min = 0,
 	max = 250, // Google's UI in the MC shows a maximum limit of 250 days, though the API doesn’t appear to have any such restriction.
@@ -106,4 +106,4 @@ const Stepper = ( {
 	);
 };
 
-export default Stepper;
+export default TimeStepper;
