+++ js/src/components/gtin-migration-banner/gtin-migration-banner.scss
